Moved hooks to src folder
