+++ src/hooks/useMe.ts
@@ -1,0 +1,6 @@
+import { useMeQuery } from "~/apollo/generated/graphql";
+
+export function useMe() {
+  const res = useMeQuery();
+  return { me: res.data?.me, ...res };
+}
